update Columns component
